⌘ 􀅇 no longer changes the focus, more accurate hit test for ring controls, move multiple does not retain multiple grabs, change color ⇨ black text

+++ Zones/sources/views/big/ZEssayView.swift
@@ -32,13 +32,21 @@ class ZEssayView: ZTextView, ZTextViewDelegate {
 	func done() { save(); exit() }
 
 	func exit() {
-		if  let e = gCurrentEssay,
-			e.lastTextIsDefault,
-			e.autoDelete {
-			e.delete()
+		if  let e = gCurrentEssay {
+			if  e.lastTextIsDefault,
+				e.autoDelete {
+				e.delete()
+			}
+
+			if  let idea = e.zone {
+				gHere = idea
+
+				gFocusRing.push()
+			}
 		}
 
 		gControllers.swapGraphAndEssay()
+		signalRegarding(.eRelayout)
 	}
 
 	func save() {
